app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
decorator==5.1.1
Django==4.0.6
frozenlist==1.3.0
future==0.18.2
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
python-telegram-bot==12.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sheetfu==1.6.1
six==1.16.0
sqlparse==0.4.2
# telegram.py==0.3.1
tornado==6.2
uritemplate==3.0.1
urllib3==1.26.10
xmltodict==0.13.0
yarl==1.7.2
